+++ Cargo.toml
@@ -2,6 +2,9 @@
 name = "ptrs"
 version = "0.1.0"
 edition = "2021"
+license = "MIT OR Apache-2.0"
+description = "A library supporting construction and integration of pluggable network transports in rust."
+
 
 [lib]
 name = "ptrs"
